CHNG: script rewritten;removed code repatition, multiple api calls, Check desc
Previously the script was not extensible as using case statements and not using functions now it is replaced with a single funciton which just pick any random repo from the repo array and get list of files and downlaod that file. Now we can just add more repo to the array and don't need to alter other code.